requirements.txt
notion-client==2.2.1
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pytz==2024.1
